7_twoknights done

+++ introduction/7_twoknights.cpp
@@ -2,12 +2,12 @@
 using namespace std;
 
 int main() {
-  int n;
+  long long n;
   cin >> n;
 
-  for (int i = 1; i <= n; i++) {
-    int totalCombinations = ((i * i) * ((i * i) - 1)) / 2;
-    int attackingWays = 4 * (i - 1) * (i - 2);
+  for (long long i = 1; i <= n; i++) {
+    long long totalCombinations = ((i * i) * ((i * i) - 1)) / 2;
+    long long attackingWays = 4 * (i - 1) * (i - 2);
     cout << totalCombinations - attackingWays << endl;
   }
 }
